+++ ikarus/controlroutines/pathfollowing.hh
@@ -12,12 +12,14 @@
 
 #include <ikarus/controlroutines/adaptivestepsizing.hh>
 #include <ikarus/controlroutines/controlinfos.hh>
+#include <ikarus/controlroutines/controlroutinebase.hh>
+#include <ikarus/controlroutines/controlroutinefactory.hh>
 #include <ikarus/controlroutines/pathfollowingfunctions.hh>
 #include <ikarus/solver/nonlinearsolver/newtonraphsonwithscalarsubsidiaryfunction.hh>
 #include <ikarus/solver/nonlinearsolver/nonlinearsolverfactory.hh>
 #include <ikarus/utils/differentiablefunctionfactory.hh>
-#include <ikarus/utils/observer/observer.hh>
-#include <ikarus/utils/observer/observermessages.hh>
+#include <ikarus/utils/broadcaster/broadcastermessages.hh>
+#include <ikarus/utils/nonlinopfactory.hh>
 
 namespace Ikarus {
 
@@ -45,6 +47,59 @@ namespace Impl {
   }
 
 } // namespace Impl
+
+template <typename NLS, typename PF, typename ASS>
+requires(Impl::checkPathFollowingTemplates<NLS, PF, ASS>())
+class PathFollowing;
+
+/**
+ * \struct PathFollowingConfig
+ * \brief Config for the Path-Following control routine
+ *
+ * \tparam PF_ the type of PathFollowing that is used (defaults to ArcLength)
+ * \tparam ASS_ the type of AdaptiveStepSizing that is used (defaults to NoOp)
+ */
+template <typename PF_ = ArcLength, typename ASS_ = AdaptiveStepSizing::NoOp>
+struct PathFollowingConfig
+{
+  using PF  = PF_;
+  using ASS = ASS_;
+
+  int steps{};
+  double stepSize{};
+  PF pathFollowingFunction{};
+  ASS adaptiveStepSizingFunction{};
+};
+
+#ifndef DOXYGEN
+PathFollowingConfig(int, double) -> PathFollowingConfig<>;
+
+template <typename PF>
+PathFollowingConfig(int, double, PF) -> PathFollowingConfig<PF>;
+
+template <typename PF, typename ASS>
+PathFollowingConfig(int, double, PF, ASS) -> PathFollowingConfig<PF, ASS>;
+
+template <typename ASS>
+PathFollowingConfig(int, double, ArcLength, ASS) -> PathFollowingConfig<ArcLength, ASS>;
+#endif
+
+/**
+ * \brief Function to create a path following instance
+ *
+ * \tparam NLS Type of the nonlinear solver
+ * \tparam PFConfig  the provided config for the path following
+ * \param config the provided config for the path following
+ * \param nonlinearSolver the provided nonlinearsolver
+ * \return PathFollowing
+ */
+template <typename NLS, typename PFConfig>
+requires traits::isSpecialization<PathFollowingConfig, std::remove_cvref_t<PFConfig>>::value
+auto createControlRoutine(PFConfig&& config, NLS&& nonlinearSolver) {
+  return PathFollowing<typename std::remove_cvref_t<NLS>::element_type, typename PFConfig::PF, typename PFConfig::ASS>(
+      std::forward<NLS>(nonlinearSolver), config.steps, config.stepSize, config.pathFollowingFunction,
+      config.adaptiveStepSizingFunction);
+}
 
 /**
  * \class PathFollowing
@@ -77,7 +132,8 @@ namespace Impl {
  */
 template <typename NLS, typename PF = ArcLength, typename ASS = AdaptiveStepSizing::NoOp>
 requires(Impl::checkPathFollowingTemplates<NLS, PF, ASS>())
-class PathFollowing : public IObservable<ControlMessages>
+class PathFollowing : public ControlRoutineBase<typename NLS::NonLinearOperator>
+
 {
 public:
   /** \brief The name of the PathFollowing method. */
@@ -118,6 +174,7 @@ private:
   double stepSize_;
   PF pathFollowingType_;
   ASS adaptiveStepSizing_;
+  SubsidiaryArgs subsidiaryArgs_;
 };
 
 } // namespace Ikarus