+++ ikarus/solver/nonlinearsolver/solverinfos.hh
@@ -8,6 +8,8 @@
 
 #pragma once
 
+#include <limits>
+
 namespace Ikarus {
 /**
  * \brief Information about the result of a non-linear solver.