+++ tests/src/testpathfollowing.cpp
@@ -15,6 +15,9 @@
 #include <ikarus/utils/init.hh>
 #include <ikarus/utils/observer/controllogger.hh>
 #include <ikarus/utils/observer/nonlinearsolverlogger.hh>
+#include <ikarus/utils/listener/controllogger.hh>
+#include <ikarus/utils/listener/nonlinearsolverlogger.hh>
+#include <ikarus/utils/nonlinearoperator.hh>
 
 using namespace Ikarus::Concepts;
 using Dune::TestSuite;
@@ -46,16 +49,15 @@ static auto simple2DOperatorArcLengthTest(DifferentiableFunction& f, typename Di
   auto linSolver  = Ikarus::LinearSolver(Ikarus::SolverTypeTag::d_LDLT);
   auto pft        = Ikarus::ArcLength{}; // Type of path following technique
 
-  auto nrSettings = Ikarus::NewtonRaphsonWithSubsidiaryFunctionConfig<decltype(linSolver)>{.linearSolver = linSolver};
-  auto nr         = Ikarus::createNonlinearSolver(nrSettings, f);
+  auto nrSettings = Ikarus::NewtonRaphsonWithSubsidiaryFunctionConfig({}, linSolver);
+  auto nr         = Ikarus::createNonlinearSolver(nrSettings, nonLinOp);
 
   auto alc = Ikarus::PathFollowing(nr, loadSteps, stepSize, pft);
 
-  auto nonLinearSolverObserver = std::make_shared<Ikarus::NonLinearSolverLogger>();
-  auto pathFollowingObserver   = std::make_shared<Ikarus::ControlLogger>();
-  nr->subscribeAll(nonLinearSolverObserver);
-  alc.subscribeAll(pathFollowingObserver);
-  const auto controlInfo              = alc.run(req);
+  auto nonLinearSolverObserver = Ikarus::NonLinearSolverLogger().subscribeTo(nr);
+  auto pathFollowingObserver   = Ikarus::ControlLogger().subscribeTo(alc);
+
+  const auto controlInfo              = alc.run();
   std::vector<int> expectedIterations = {1, 3, 3, 3, 3};
   Eigen::Vector2d expectedDisplacement;
   expectedDisplacement << 0.0883524725970593, 0.3486891582376427;
@@ -107,14 +109,13 @@ static auto simple2DOperatorLoadControlTest(DifferentiableFunction& f, typename 
   auto linSolver  = Ikarus::LinearSolver(Ikarus::SolverTypeTag::d_LDLT);
   auto pft        = Ikarus::LoadControlSubsidiaryFunction{}; // Type of path following technique
 
-  auto nrSettings = Ikarus::NewtonRaphsonWithSubsidiaryFunctionConfig<decltype(linSolver)>{.linearSolver = linSolver};
-  auto nr         = Ikarus::createNonlinearSolver(nrSettings, f);
-  auto lc         = Ikarus::PathFollowing(nr, loadSteps, stepSize, pft);
-  auto nonLinearSolverObserver = std::make_shared<Ikarus::NonLinearSolverLogger>();
-  auto pathFollowingObserver   = std::make_shared<Ikarus::ControlLogger>();
-  nr->subscribeAll(nonLinearSolverObserver);
-  lc.subscribeAll(pathFollowingObserver);
-  const auto controlInfo              = lc.run(req);
+  auto nrSettings              = Ikarus::NewtonRaphsonWithSubsidiaryFunctionConfig({}, linSolver);
+  auto nr                      = Ikarus::createNonlinearSolver(nrSettings, nonLinOp);
+  auto lc                      = Ikarus::PathFollowing(nr, loadSteps, stepSize, pft);
+  auto nonLinearSolverObserver = Ikarus::NonLinearSolverLogger().subscribeTo(nr);
+  auto pathFollowingObserver   = Ikarus::ControlLogger().subscribeTo(lc);
+
+  const auto controlInfo              = lc.run();
   std::vector<int> expectedIterations = {2, 3, 3, 3, 3};
   Eigen::Vector2d expectedDisplacement;
   expectedDisplacement << 0.0908533884835060, 0.3581294588381901;
@@ -140,14 +141,13 @@ static auto simple2DOperatorDisplacementControlTest(DifferentiableFunction& f,
 
   auto pft = Ikarus::DisplacementControl{controlledIndices}; // Type of path following technique
 
-  auto nrSettings = Ikarus::NewtonRaphsonWithSubsidiaryFunctionConfig<decltype(linSolver)>{.linearSolver = linSolver};
-  auto nr         = Ikarus::createNonlinearSolver(nrSettings, f);
-  auto dc         = Ikarus::PathFollowing(nr, loadSteps, stepSize, pft);
-  auto nonLinearSolverObserver = std::make_shared<Ikarus::NonLinearSolverLogger>();
-  auto pathFollowingObserver   = std::make_shared<Ikarus::ControlLogger>();
-  nr->subscribeAll(nonLinearSolverObserver);
-  dc.subscribeAll(pathFollowingObserver);
-  const auto controlInfo              = dc.run(req);
+  auto nrSettings              = Ikarus::NewtonRaphsonWithSubsidiaryFunctionConfig({}, linSolver);
+  auto nr                      = Ikarus::createNonlinearSolver(nrSettings, nonLinOp);
+  auto dc                      = Ikarus::PathFollowing(nr, loadSteps, stepSize, pft);
+  auto nonLinearSolverObserver = Ikarus::NonLinearSolverLogger().subscribeTo(nr);
+  auto pathFollowingObserver   = Ikarus::ControlLogger().subscribeTo(dc);
+
+  const auto controlInfo              = dc.run();
   std::vector<int> expectedIterations = {3, 3, 3, 3, 3};
   Eigen::Vector2d expectedDisplacement;
   expectedDisplacement << 0.5, 1.4781013410920430;